Move address status to its own file
